+++ src/net/sf/persism/SQL.java
@@ -6,10 +6,8 @@ import java.util.regex.Pattern;
 /**
  * Simple wrapper for SQL String. Mainly to allow for overloads to fetch/query methods.
  *
- * @see <a href="https://sproket.github.io/Persism/manual.html">Using the new Query/Fetch methods</a>
+ * @see <a href="https://sproket.github.io/Persism/manual2.html">Using the new Query/Fetch methods</a>
  */
-// todo wrong link for now....
-// todo cache these if we do any parsing so we only parse once.
 public final class SQL {
 
     final String sql;